# [Data Normalization] Improve performance