to: @ryandeivert @austinbyers 
cc: @airbnb/streamalert-maintainers
size: median
resolves #717 

## Background
In current Data Normalization implementation, Data Normalization is applied to a record right before rule matching. So one record may be normalized multiple times when there are different rules using different `datatypes`. For example,
* Rule 1 
```
@rule(datatypes=['command'], ...)
def rule_get_commands(rec):
...
```
* Rule 2
```
@rule(datatypes=['sourceAddress'], ...)
def rule_get_ip_address(rec):
...
```
* Rule 3
```
@rule(datatypes=['command', `sourceAddress`], ...)
def rule_get_command_and_ip_address(rec):
...
```
In this case, one record will be normalized 3 times!!! The performance will become bottleneck when you will thousands records read from files from S3 bucket, and it will result lambda function to be timed out.

## Changes
* In rules engine, aggregate datatypes from all rules and normalize records only once before any rules matching.
* This change also update `get_keys` helper to skip normalization information.

## Testing
The testing environment has multiple rules using data normalization with different datatypes. And running time is based on `timeit`.
* 2761 carbonblack netconn records read from S3 file 
  * Before improvement, the running time is 24s
  * Apply improvement, the running time is 10s
  
* 7713 carbonblack event records read from S3 file
  * Before improvement, the running time is 37s
  * Apply improvement, the running time is 25s
  
* 1521 cloudtrail records read from S3 file
  * Before improvement, the running time is 584s
  * Apply improvement, the running time is 3s

Cloudtrail records seem to have significant benefit because they contains nested keys and ip addresses which will be normalized in my test environment.